move ViewNextCommand and ViewCurrentCommand to commands package
