+++ src/main/java/seedu/address/logic/commands/ViewCurrentCommand.java
@@ -1,21 +1,22 @@
-package seedu.address.logic;
+package seedu.address.logic.commands;
 
 import seedu.address.commons.core.UiCommandInteraction;
+import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 
 import static java.util.Objects.requireNonNull;
 
-public class ViewNextCommand extends Command {
+public class ViewCurrentCommand extends Command {
 
-    public static final String COMMAND_WORD = "viewNext";
+    public static final String COMMAND_WORD = "viewCurrent";
 
-    public static final String MESSAGE_SUCCESS = "Listed Next Month Duty";
+    public static final String MESSAGE_SUCCESS = "Listed Current Month Duty";
 
     public CommandResult execute(Model model, CommandHistory history) {
         requireNonNull(model);
-        return new CommandResult(MESSAGE_SUCCESS, UiCommandInteraction.CALENDAR_NEXT);
+        return new CommandResult(MESSAGE_SUCCESS, UiCommandInteraction.CALENDAR_CURRENT);
     }
 
     @Override
